[Master] Decouple Master and Master extensions
This is to continue the work on the old repo. In particular, this patch
decouples the extensions (e.g., MongoDB module) from Master so that
these modules can extend Master in a non-intrusive way.
